:cookie: split js into modules
